+++ CHANGELOG.md
@@ -43,6 +43,7 @@ in 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
   * 2025-08-27
     * Finalized translations
+    * Added width oversize attribute for edge banding material type
   * 2025-08-24
     * Updated RU and ZH languages
   * 2025-08-23